requirement.txt
SpeechRecognition
pydub

PyAudio
requests

python-dotenv
pydantic

google-genai
